Rename path management file to better name
