+++ src/integrations/pathManager.ts
